adds ability to set data path
previously, we had always downloaded data into `SampleData/DownloadedData` within the current directory.

that's bad on the cluster because we want code to live in /home, and data on /scratch. so this allows user to set path as a command line argument and, if `module list` runs without an error, assumes we should put it in /scratch.